requirements.txt
# School Attendance Management System - Requirements
# Install with: pip install -r requirements.txt

# Computer Vision & Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
imutils==0.5.4

# Machine Learning & Data
scikit-learn==1.3.1
pandas==2.0.3

# Database
mysql-connector-python==8.1.0

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2

# Background Scheduling
APScheduler==3.10.4

# SMS Notifications (Optional)
twilio==8.10.0
requests==2.31.0

# PDF Generation & Reporting (Optional)
reportlab==4.0.7
PyPDF2==3.0.1

# Data Validation
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0


